Seeding database

+++ eShopSolution.Data/Entities/Category.cs
@@ -8,7 +8,7 @@ namespace eShopSolution.Data.Entities
     public class Category
     {
         public int Id { get; set; }
-        public decimal SortOrder { get; set; }
+        public int SortOrder { get; set; }
         public bool IsShowOnHome { get; set; }
         public int? ParentId { get; set; }
         public Status Status { get; set; }
